network: reconstruct net_chan.c/net_ws.c against the binary's stub network stack

+++ src/engine/client.h
@@ -248,11 +248,11 @@ typedef struct
 {
 	// generated on client side
 	usercmd_t	cmd;		// cmd that generated the frame
-	double		senttime;	// time cmd was sent off
+	float		senttime;	// time cmd was sent off
 	int			delta_sequence;		// sequence number to delta from, -1 = full update
 
 	// received from server
-	double		receivedtime;	// time message was received, or -1
+	float		receivedtime;	// time message was received, or -1
 	qboolean	invalid;		// true if the packet_entities delta was invalid
 
 	player_state_t	playerstate[MAX_CLIENTS];	// message received that reflects performing the usercmd
